rename some traits
